+++ NEWS.md
@@ -4,6 +4,17 @@
 
 - `Axis` does now accept both a `Bool` and a `Tuple{Bool, Bool}` as values for `xtrimspine` and `ytrimspine` to trim only one end of the spine [#2171](https://github.com/JuliaPlots/Makie.jl/pull/2171).
 
+## v0.17.13
+
+- Fix boundingboxes [#2184](https://github.com/JuliaPlots/Makie.jl/pull/2184).
+- Fix highclip/lowclip in meshscatter, poly, contourf, barplot [#2183](https://github.com/JuliaPlots/Makie.jl/pull/2183).
+- fix gridline updates [#2196](https://github.com/JuliaPlots/Makie.jl/pull/2196).
+- fix glDisablei argument order, which crashed some Intel drivers
+
+## v0.17.12
+
+- Fix stackoverflow in show [#2167](https://github.com/JuliaPlots/Makie.jl/pull/2167)
+
 ## v0.17.11
 
 - `rainclouds`(!) now supports `violin_limits` keyword argument, serving the same.